Pass CreateDomain test
Also add TODO about some packages to glide.yaml

Add teardown code taht clears the `domains` database

CreateDomain relies on GetDomain in order to verify uniqueness of the name of the domain to be made. This is an alternative to relying on the error message of the underlying SQL driver.